+++ core/events_types.py
@@ -1,0 +1,76 @@
+# This file is part of project Sverchok. It's copyrighted by the contributors
+# recorded in the version control history of the file, available from
+# its original location https://github.com/nortikin/sverchok/commit/master
+#
+# SPDX-License-Identifier: GPL3
+# License-Filename: LICENSE
+
+
+from enum import Enum, auto
+
+
+class BlenderEventsTypes(Enum):
+    # Enumeration of signals which Blender provide
+    tree_update = auto()  # this updates is calling last with exception of creating new node
+    monad_tree_update = auto()
+    node_update = auto()  # it can be called last during creation new node event
+    add_node = auto()   # it is called first in update wave
+    copy_node = auto()  # it is called first in update wave
+    free_node = auto()  # it is called first in update wave
+    # all properties of free node are set to default and can't be change, so it loos its ID
+    # never the less it keeps its actual name even if the node was renamed
+    add_link_to_node = auto()  # it can detects only manually created links
+    # link leeds to wrong memory address out of the signal method
+    # using sverchok defined properties (link_id) leads to crash during the event
+    node_property_update = auto()  # can be in correct in current implementation
+    undo = auto()  # changes in tree does not call any other update events
+    frame_change = auto()
+
+    def print(self, updated_element=None):
+        event_name = f"EVENT: {self.name: <30}"
+        if updated_element is not None:
+            element_data = f"IN: {updated_element.bl_idname: <25} INSTANCE: {updated_element.name: <25}"
+        else:
+            element_data = ""
+        print(event_name + element_data)
+
+
+class SverchokEventsTypes(Enum):
+    # Enumeration of events which can be translate from Blender signals for Sverchok workflow
+    add_node = auto()
+    copy_node = auto()
+    free_node = auto()
+    node_property_update = auto()
+    add_link = auto()
+    free_link = auto()
+    undo = auto()
+    undefined = auto()  # probably should be deleted
+    frame_change = auto()
+
+    def print(self, updated_element=None):
+        event_name = f"SVERCHOK EVENT: {self.name: <21}"
+        element_data = f"NODE: {updated_element.name}" if updated_element else ""
+        print(event_name + element_data)
+
+
+EVENT_CONVERSION = {
+    BlenderEventsTypes.tree_update: SverchokEventsTypes.undefined,
+    BlenderEventsTypes.monad_tree_update: SverchokEventsTypes.undefined,
+    BlenderEventsTypes.node_update: SverchokEventsTypes.undefined,
+    BlenderEventsTypes.add_node: SverchokEventsTypes.add_node,
+    BlenderEventsTypes.copy_node: SverchokEventsTypes.copy_node,
+    BlenderEventsTypes.free_node: SverchokEventsTypes.free_node,
+    BlenderEventsTypes.add_link_to_node: SverchokEventsTypes.add_link,
+    BlenderEventsTypes.node_property_update: SverchokEventsTypes.node_property_update,
+    BlenderEventsTypes.undo: SverchokEventsTypes.undo,
+    BlenderEventsTypes.frame_change: SverchokEventsTypes.frame_change
+}
+
+
+IS_WAVE_END = {
+    BlenderEventsTypes.tree_update: True,
+    BlenderEventsTypes.node_property_update: True,
+    BlenderEventsTypes.monad_tree_update: True,
+    BlenderEventsTypes.undo: True,
+    BlenderEventsTypes.frame_change: True
+}
